// aircraft-maintenance-ui/src/app/services/models/page-response-scheduled-check-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ScheduledCheckDto } from '../models/scheduled-check-dto';
export interface PageResponseScheduledCheckDto {
  content?: Array<ScheduledCheckDto>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}